refactor: unnecessary styled modal component
