+++ packages/shared/src/components/modals/DeleteAccountModal.tsx
@@ -1,6 +1,6 @@
 import React, { ReactElement, MouseEvent, useState } from 'react';
 import { Button } from '../buttons/Button';
-import { ModalProps } from './StyledModal';
+import { ModalProps } from './common/Modal';
 import {
   ConfirmationModal,
   ConfirmationHeading,
