+++ packages/shared/src/components/modals/PostModal.tsx
@@ -9,7 +9,7 @@ import classNames from 'classnames';
 import { useQuery } from 'react-query';
 import request from 'graphql-request';
 import { Modal } from './common/Modal';
-import { ModalProps } from './StyledModal';
+import { ModalProps } from './common/Modal';
 import { useHideOnModal } from '../../hooks/useHideOnModal';
 import { useResetScrollForResponsiveModal } from '../../hooks/useResetScrollForResponsiveModal';
 import { PostContent, SCROLL_OFFSET } from '../post/PostContent';
