+++ packages/shared/src/components/modals/ReportPostModal.tsx
@@ -2,7 +2,7 @@ import React, { ReactElement, MouseEvent, useState, useRef } from 'react';
 import { Radio } from '../fields/Radio';
 import { Post } from '../../graphql/posts';
 import { Checkbox } from '../fields/Checkbox';
-import { ModalProps } from './StyledModal';
+import { ModalProps } from './common/Modal';
 import { Button } from '../buttons/Button';
 import { PostBootData } from '../../lib/boot';
 import { Modal } from './common/Modal';
